extract login logic from login form to login page
