+++ client/src/App/components/forms/SigninForm.js
@@ -1,58 +1,22 @@
-import React, { useState, useEffect } from 'react';
-import { Redirect, withRouter } from 'react-router-dom';
-
-import Notification from '../notification/Notification';
+import React, { useState } from 'react';
 
 import './forms.scss';
 
-function SigninForm(props) {
+function SigninForm({submitData}) {
+
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
-    const [loggedIn, setLoggedIn] = useState(false);
-    const [notification, setNotification] = useState({});
-
-    const DURATION = 5000; // in sync with scss animation TO REFACTOR
-
-    useEffect(() => {
-        const {state = {}} = props.history.location;
-        const {type, message} = state;
-        if(type){
-            displayNotification(type, message);
-        };
-    }, []);
-
-    function displayNotification (type, message) {
-        setNotification({type, message, display:true})
-        setTimeout(() => setNotification({type:"", message:"", display:false}), DURATION);
-    }
 
     function submitForm(e) {
         e.preventDefault();
 
-        const fetchOpt = {
-            method: 'POST',
-            headers: {"content-type":"application/json"},
-            body: JSON.stringify({email, password})
-        };
-
-        return fetch('/auth/login', fetchOpt)
-                .then(response => response.json())
-                .then(message =>  {
-                    setEmail("");
-                    setPassword("");
-                    if(message.error || message.message.startsWith("Invalid")) {
-                       displayNotification('error','Authentication failed');
-                       return;
-                    };
-                    return setLoggedIn(true);
-                })
-                .catch(error => console.error(error));
+        submitData({email, password});
+        setEmail("");
+        setPassword("");        
     }
 
     return ( 
         <form className="form-container" onSubmit={submitForm}>
-            {loggedIn && <Redirect to={{pathname:"/", state: {type: 'success', message: 'You successfully logged in!'}}}/>}
-            {notification.type && <Notification {...notification} />}
             <div className="form-field" >
                 <label htmlFor="email">Email: </label>
                 <input 
@@ -85,4 +49,4 @@ function SigninForm(props) {
     );
 };
 
-export default withRouter(SigninForm);
+export default SigninForm;
